#235 [rename] 행동 로그 기록을 위한 엘라스틱 서치 설정을 위한 properties를 behaviorlog에서 common으로 이동한다.
